Add error handling
Added src folder
Added try/catch block for error handling
